Generalize Truetype to Sfnt, with Opentype as its sibling
`Sfnt` is the container format both TrueType and OpenType specialise, and the code
already divided along that seam: ten of the table accessors are shared — `head`,
`cmap`, `hhea`, `hmtx`, `maxp`, `post`, `os2`, `name`, the directory machinery and
everything built on it — while only `loca`, `glyf`, `glyphClosure` and `subset`
are TrueType's quadratic-outline logic. The shared members move to `trait Sfnt`;
`Truetype` keeps the four; `Opentype` is new.

`Sfnt(data)` dispatches on the table directory: a `CFF ` table means PostScript
outlines, so the font is `Opentype`. This fixes a real defect rather than just
tidying. facsimile's own comment admitted that "embedded TrueType and OpenType
programs surface as phoenicia `Ttf`s", and `FontEmbedder` then called `subset` on
them — `safely(...).or(ttf)` swallowed the failure and embedded the whole font.
A CFF font is now typed correctly, subsetting is conditional on actually having
TrueType outlines, and the fallback is explicit.

Two consequences worth recording.

Dispatching means construction must read the directory, where wrapping bytes in a
font was previously total and lazy. A test embeds deliberately-unparseable bytes,
relying on that; the dispatch is therefore tolerant — an unreadable directory
yields a `Truetype`, as before.

`Glyph` was indexed by `Truetype & Singleton` and is now indexed by `Sfnt &
Singleton`, since glyph lookup is a shared, cmap-driven operation.
